+++ include/hyperrmq/binary_tree.hpp
@@ -8,17 +8,17 @@
 namespace hyperrmq {
 
 struct BinaryTree {
-    int32_t None = -1;
+    int64_t None = -1;
 
-    int32_t n;
-    std::vector<int32_t> parent;
-    std::vector<int32_t> left;
-    std::vector<int32_t> right;
-    int32_t root;
+    int64_t n;
+    std::vector<int64_t> parent;
+    std::vector<int64_t> left;
+    std::vector<int64_t> right;
+    int64_t root;
 
     BinaryTree();
 
-    explicit BinaryTree(const std::vector<int32_t> parent_inorder);
+    explicit BinaryTree(const std::vector<int64_t> parent_inorder);
 
     std::string to_BP_string(std::string left_delim = "(",
                              std::string middle_delim = ")",
