+++ include/hyperrmq/three_level_prefix_sum.hpp
@@ -4,7 +4,7 @@
 
 namespace hyperrmq {
 
-template <typename T = uint32_t, uint32_t K = 32, uint32_t L = 32>
+template <typename T = uint64_t, uint64_t K = 32, uint64_t L = 32>
 struct ThreeLevelPrefixSum {
    public:
     ThreeLevelPrefixSum();
